+++ salve_ipc/server_functions/__init__.py
@@ -1,5 +1,6 @@
 from .autocompletions import find_autocompletions  # noqa: F401
 from .definitions import get_definition  # noqa: F401
 from .highlight import get_highlights  # noqa: F401
-from .misc import Token, generic_tokens, is_unicode_letter  # noqa: F401
+from .highlight import Token, generic_tokens  # noqa: F401
+from .misc import is_unicode_letter  # noqa: F401
 from .replacements import get_replacements  # noqa: F401
