+++ salve_ipc/server_functions/highlight/__init__.py
@@ -1,0 +1,2 @@
+from .highlight import get_highlights  # noqa: F401
+from .tokens import Token, generic_tokens  # noqa: F401
