+++ salve_ipc/server_functions/highlight/highlight.py
